@OmarManzoor 2023-07-11
> Thank you for exploring it, @OmarManzoor! 🤝
> 
> Before proceeding, I think it makes sense to profile the execution of `RadiusNeighbors.predict_proba` to see if it is worth optimizing (not to loose time as we did in the past with #2 and [scikit-learn#26316](https://github.com/scikit-learn/scikit-learn/pull/26316)).
> 
> Also you have selected your fork's `main` branch. Is it intentional?

Yes this is just so we can discuss points and potential approaches before actually creating a PR in scikit learn.