@OmarManzoor 2023-07-11
Copying the comment I posted from the ticket

Using some simple timeit statements I was able to get the following results

```
--------------------------Iteration 1--------------------------
Start predict probabilities
Time taken by radius_neighbors: 19.6450375
Time taken to compute probabilities: 36.3680472
-----------------------------------------------------------------
--------------------------Iteration 2--------------------------
Start predict probabilities
Time taken by radius_neighbors: 39.5624981
Time taken to compute probabilities: 44.52882960000001
-----------------------------------------------------------------
--------------------------Iteration 3--------------------------
Start predict probabilities
Time taken by radius_neighbors: 29.812655099999972
Time taken to compute probabilities: 43.69632340000004
-----------------------------------------------------------------
And two more iterations in which I ran the complete script again including the fit method.

Time taken by radius_neighbors: 32.9073425
Time taken to compute probabilities: 42.0797534
---------------------------------------------------------------
Start predict probabilities
Time taken by radius_neighbors: 16.559098600000002
Time taken to compute probabilities: 36.1868528
```