@jjerphan 2023-07-11
Thanks for the report. It looks like it is worth optimizing `RadiusNeighbors.predict_proba`.