@Micky774 2023-07-11
> I wasn't able to actually successfully profile using the profiler that you had suggested (py-spy).

We may want to see if we can fix whatever is going wrong with that. But ditto, all signs point to there being a clear benefit to optimizing this section.