@OmarManzoor 2023-07-12
> Yes, please reuse the asv benchmark scripts.

Yeah did that exactly but when I ran on a larger set of data, it results in memory errors like
`python(17542,0x1f03e5e00) malloc: Incorrect checksum for freed object 0x135827000: probably modified after being freed.`
`segmentation fault`