@jjerphan 2023-07-12
Could you craft a minimal reproducer?