@OmarManzoor 2023-07-12
> Could you craft a minimal reproducer?

https://gist.github.com/OmarManzoor/4ccf778046b66ddc02122d7db1f61d76