@OmarManzoor 2023-07-13
> You can try to use `asan` if you are using `clang` otherwise you can debug using `gdb --args python script.py`. I recommend forcing using a single thread to debug (this can be done with `taskset(1)` on Linux).

I am on on Mac M1 now.