@OmarManzoor 2023-07-13
```console
· Running 2 total benchmarks (2 commits * 1 environments * 1 benchmarks)
[ 0.00%] · For scikit-learn commit 0ad31271 <radius_neighbors_predict_pwdb> (round 1/1):
[ 0.00%] ·· Benchmarking conda-py3.11-cython-joblib-numpy-pandas-scipy-threadpoolctl
[50.00%] ··· radius_neighbors_predict.BruteRadiusNeighborsClass.time_predict_proba                                                                                  ok
[50.00%] ··· ========= ============
             --        n_test / n_features
             --------- ------------
              n_train   1000 / 100 
             ========= ============
               10000     195±20ms  
             ========= ============

[50.00%] · For scikit-learn commit dfd299b1 (round 1/1):
[50.00%] ·· Building for conda-py3.11-cython-joblib-numpy-pandas-scipy-threadpoolctl.
[50.00%] ·· Benchmarking conda-py3.11-cython-joblib-numpy-pandas-scipy-threadpoolctl
[100.00%] ··· radius_neighbors_predict.BruteRadiusNeighborsClass.time_predict_proba                                                                                  ok
[100.00%] ··· ========= ============
              --        n_test / n_features
              --------- ------------
               n_train   1000 / 100 
              ========= ============
                10000     226±20ms  
              ========= ============

| Change   | Before [dfd299b1]    | After [0ad31271] <radius_neighbors_predict_pwdb>   |   Ratio | Benchmark (Parameter)                                                                   |
|----------|----------------------|----------------------------------------------------|---------|-----------------------------------------------------------------------------------------|
| -        | 226±20ms             | 195±20ms                                           |    0.86 | radius_neighbors_predict.BruteRadiusNeighborsClass.time_predict_proba(10000, 1000, 100) |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.

```

ASV Benchmark: https://gist.github.com/OmarManzoor/444ee05b9c902ca795fc7e377f4d5714